wandboard: fix mipi-csi exclusion and reorg
Fix a small bug when excluding MIPI from the build configuration.  There
was a '2' missing at the end of the CONFIG_IMX_HAVE_PLATFORM_IMX_MIPI_CSI2
var.

Also, reorganized the code so that all of the IPU capture, MIPI, and sensor
init is called using wand_init_mipi_csi(), which is a stub now when MIPI
is deslected.