ARM: clk-imx6q: fix video divider for revision 1.0 of i.MX6q
As post dividers do not work on i.MX6Q revision 1.0 they must be fixed
to 1. As the table index was wrong, a divider a of 4 could still be
requested which implied the clock not to be set properly. This is the
root cause of the HDMI not working at high resolution on T1.0 version of
the SoC, giving the following error:
	mxc_sdc_fb fb.27: timeout when waiting for flip irq

See this thread for details:
	https://lists.yoctoproject.org/pipermail/meta-freescale/2014-December/thread.html#11776

Signed-off-by: Gary Bisson <bisson.gary@gmail.com>

(From boundary devices' Eric Nelson)